# public/swagger.yaml
openapi: 3.0.3
info:
  title: Air5Star eCommerce AI-Powered API
  description: |
    AI-powered features API for Air5Star HVAC eCommerce platform.
    
    This API provides intelligent features including:
    - Smart chatbot integration with product lookup
    - Advanced search with autocomplete and highlighting
    - Personalized product recommendations
    - Abandoned cart email automation
    - Inventory forecasting and alerts
    
    ## Security
    - All endpoints use HTTPS
    - API keys stored in environment variables
    - HMAC signature verification for webhooks
    - Minimal PII exposure to third-party services
    
    ## Rate Limiting
    - Search endpoints: 100 requests/minute
    - Recommendation endpoints: 50 requests/minute
    - Webhook endpoints: 1000 requests/hour
  version: 1.0.0
  contact:
    name: Air5Star Development Team
    email: dev@air5star.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://air5star.com/api
    description: Production server
paths:
...
  /inventory-forecast:
    get:
      tags:
        - Inventory Management
      summary: Get inventory forecast and alerts
      description: |
        Generate inventory forecasts, reorder alerts, and demand predictions
        based on historical sales data and seasonal patterns.
      parameters:
        - name: product_id
          in: query
          description: Specific product ID to forecast
          schema:
            type: string
          example: "1"
        - name: category
          in: query
          description: Product category to forecast
          schema:
            type: string
          example: "Air Conditioners"
        - name: days_ahead
          in: query
          description: Number of days to forecast
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
        - name: include_seasonality
          in: query
          description: Include seasonal factors in forecast
          schema:
            type: boolean
            default: true
        - name: confidence_level
          in: query
          description: Forecast confidence level
          schema:
            type: number
            minimum: 0.5
            maximum: 0.99
            default: 0.95
      responses:
        '200':
          description: Forecast results and alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryForecastResponse'
              example:
                success: true
                data:
                  forecasts:
                    - product_id: "1"
                      product_name: "Split AC 1.5 Ton"
                      category: "Air Conditioners"
                      current_stock: 25
                      predicted_demand: 18
                      reorder_point: 15
                      suggested_order_quantity: 30
                      forecast_accuracy: 0.85
                      risk_level: "medium"
                      seasonal_factors:
                        - month: 4
                          factor: 1.5
                        - month: 5
                          factor: 2.0
                      trend_analysis:
                        direction: "increasing"
                        strength: 0.7
                  alerts:
                    - product_id: "3"
                      product_name: "Window AC 1 Ton"
                      current_stock: 8
                      reorder_point: 15
                      days_until_stockout: 12
                      priority: "high"
                      suggested_action: "Reorder immediately"
                  summary:
                    total_products: 5
                    high_risk_products: 2
                    urgent_alerts: 1
                    forecast_period_days: 30
                    generated_at: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    
    post:
      tags:
        - Inventory Management
      summary: Upload sales data for forecasting
      description: |
        Upload historical sales data to improve forecast accuracy.
        Accepts bulk sales data in JSON format.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SalesDataPoint'
            example:
              - product_id: "1"
                date: "2024-01-01T00:00:00Z"
                quantity_sold: 5
                revenue: 175000
                category: "Air Conditioners"
              - product_id: "1"
                date: "2024-01-02T00:00:00Z"
                quantity_sold: 3
                revenue: 105000
                category: "Air Conditioners"
      responses:
        '200':
          description: Sales data uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      records_processed:
                        type: integer
                      date_range:
                        type: object
                        properties:
                          from:
                            type: integer
                            format: int64
                          to:
                            type: integer
                            format: int64
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    
    put:
      tags:
        - Inventory Management
      summary: Export sales data
      description: |
        Export sales data in JSON or CSV format with optional filtering.
      parameters:
        - name: format
          in: query
          description: Export format
          schema:
            type: string
            enum: ["json", "csv"]
            default: "json"
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: start_date
          in: query
          description: Start date filter
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: End date filter
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Sales data exported
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SalesDataPoint'
                  summary:
                    type: object
                    properties:
                      total_records:
                        type: integer
                      total_revenue:
                        type: number
                      total_quantity:
                        type: integer
            text/csv:
              schema:
                type: string
                example: |
                  "Product ID","Date","Quantity Sold","Revenue","Category"
                  "1","2024-01-01T00:00:00Z","5","175000","Air Conditioners"
        '500':
          $ref: '#/components/responses/InternalError'
    
    head:
      tags:
        - Inventory Management
      summary: Health check for inventory service
      responses:
        '200':
          description: Service is healthy
...
components:
  securitySchemes:
    HmacSignature:
      type: apiKey
      in: header
      name: X-Tidio-Signature
      description: HMAC SHA-256 signature for webhook verification
  schemas:
...
    InventoryForecastResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            forecasts:
              type: array
              items:
                $ref: '#/components/schemas/ForecastResult'
            alerts:
              type: array
              items:
                $ref: '#/components/schemas/InventoryAlert'
            summary:
              type: object
              properties:
                total_products:
                  type: integer
                high_risk_products:
                  type: integer
                urgent_alerts:
                  type: integer
                forecast_period_days:
                  type: integer
                generated_at:
                  type: string
                  format: date-time
    ForecastResult:
      type: object
      properties:
        product_id:
          type: string
        product_name:
          type: string
        category:
          type: string
        current_stock:
          type: integer
        predicted_demand:
          type: integer
        reorder_point:
          type: integer
        suggested_order_quantity:
          type: integer
        forecast_accuracy:
          type: number
          minimum: 0
          maximum: 1
        risk_level:
          type: string
          enum: ["low", "medium", "high"]
        seasonal_factors:
          type: array
          items:
            type: object
            properties:
              month:
                type: integer
                minimum: 1
                maximum: 12
              factor:
                type: number
        trend_analysis:
          type: object
          properties:
            direction:
              type: string
              enum: ["increasing", "decreasing", "stable"]
            strength:
              type: number
              minimum: 0
              maximum: 1
    InventoryAlert:
      type: object
      properties:
        product_id:
          type: string
        product_name:
          type: string
        current_stock:
          type: integer
        reorder_point:
          type: integer
        days_until_stockout:
          type: integer
        priority:
          type: string
          enum: ["urgent", "high", "medium", "low"]
        suggested_action:
          type: string
    SalesDataPoint:
      type: object
      required:
        - product_id
        - date
        - quantity_sold
        - revenue
      properties:
        product_id:
          type: string
        date:
          type: string
          format: date-time
        quantity_sold:
          type: integer
          minimum: 0
        revenue:
          type: number
          minimum: 0
        category:
          type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid request parameters"
            details:
              - field: "email"
                message: "Invalid email format"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Internal server error"
...